Use Kubernetes services account if possible
If fluentd is running in a pod, it has sufficient access to the
Kubernetes API to enrich log records with pod and namespace metadata.
When URL / token / etc. are not implicitly specified, try and grab the
service account's credentials from the well-known directory.

Bonus: Fix version of fluentd gem required for testing.

Issue #12