thread safety fix: do not mutate cached object
By setting 'container_name' on the metadata object returned by
`@cache.getset`, we create a race condition where containers
in the same pod will end up with an incorrect 'container_name'
assigned.

In the old code metadata['kubernetes'] was effectively shared
for all records within a pod, so the 'container_name' that ends
up in the final record was determined by whichever record happened
to set 'container_name' last.

Instead, merge the cached object into the metadata that we create
for each record.