deep copy metadata in filter_stream_from_files
This change makes deep copy of the metadata, along with kubernetes
and docker fields, to ensure they are unique per record and changeable.

As currently implemented, filter_stream_from_files method creates
a Hash 'metadata' for multiple records simultaneously and Hash::merge
does only a shallow copy of the Hash. If any other filter in
the pipeline then attempts to change anything in the 'kubernetes' or
'docker' fields, change gets propagated to the related records
sharing fluentd tag.