Turn off JSON payload merges by default
Merging JSON payloads into the log metadata by default is not a good
idea because of the potential negative impact it can have on the fields
for Elasticsearch.

For example, imagine each log line emitted by a container looks like
the following:
```
{ "abca072a2mwqpc78": 0 }
```
Where the key is "`abca072a2mwqpc78`" and its value is "`0`".

If the container generates a random value for that key, or perhaps a
number in a monotonically increasing sequence, thousands of fields can
be generated for the combined set of JSON documents indexed into
Elasticsearch.  Without an additional filter that could detect such bad
behaviors, and compare against the existing mapping of a target index,
this can cause major problems for Elasticsearch.  Each new field added
to a mapping in Elasticsearch causes a cluster state transition, which
are very very slow (on the order of seconds), and an commensurate
increase in memory usage to track those fields (which can cause
Elasticsearch to exhaust its HEAP, or at least engage in heavy garbage
collection processing).

Use of this option should only be done if the user knows well what they
are doing.

Additionally, with this commit we are deprecated the use of the option,
and will be removing the option and its functionality in a future
release.  Please consider the use of the fluentd filter_parser [1]
instead.

[1] https://docs.fluentd.org/v0.12/articles/filter_parser