get rid of dependency on activesupport
Some transitive dependency on some newer project has crept in
via activesupport.  We only use activesupport for a couple of
trivial methods, so just implement those methods in our code.
Due to needing to support ruby 2.0.0, we have to lock down the
versions of several of the components we use.