use kubeclient 4.4.latest; refactor tests for k8s api discovery
Update to kubeclient 4.4.latest - make dependency now on versions
between 4 and 5.  The new api makes a discovery call to
/api/v1 - I had to refactor the tests to take that into consideration.
I've tested this with fluentd with ocp 4.2 logging - tests pass.
I'm bumping the version to 2.3.0 from 2.2.x due to the big change
in the kubeclient version.  I also got rid of the dependency on
activesupport.